requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==3.2.16
django-allauth==0.54.0
psycopg2-binary==2.9.5
python-decouple==3.6
pytz==2022.5
sqlparse==0.4.3
pillow
python-dotenv
django-storages==1.14.2
boto3==1.28.15
gunicorn==21.2.0
@vercel/blob
